Refactored Program into more classes and functions, will need to add comments later.
