+++ NetBeans Project/CS321MagicDeckBuilder/src/mtg_data/deck_1698812135250.json
@@ -73,12 +73,12 @@
             "cardID": 461187
         },
         {
-            "quantity": 3,
-            "cardID": 460928
+            "quantity": 2,
+            "cardID": 460951
         },
         {
             "quantity": 2,
-            "cardID": 460951
+            "cardID": 460928
         }
     ],
     "name": "Deck Name",
